Refactor Validator tests to use common test subclass
