+++ projects/hipblaslt/tensilelite/Tensile/Tests/unit/test_ValidateNglAndNll.py
@@ -20,21 +20,18 @@
 # CTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 ################################################################################
 
-import unittest
 from rocisa.instruction import SWaitCnt, SBarrier
 
-from Tensile.Components.CustomSchedule import ScheduleInfo
 from Tensile.Components.CMSValidator import verify_lrs_and_grs
-from test_CustomSchedule import create_base_kernel
+from cms_validation_base import CMSValidationTestBase
 
-class TestValidateNglAndNll(unittest.TestCase):
-    def setUp(self):
-        self.kernel = create_base_kernel()
-        self.num_vmfma = 2 * self.kernel["MIWaveTileA"] * self.kernel["MIWaveTileB"]
+class TestValidateNglAndNll(CMSValidationTestBase):
+    def validation_function(self, sched, kernel_dict, codePathIdx):
+        return verify_lrs_and_grs(sched, kernel_dict, codePathIdx)
 
-    def make_simple_schedule(self, shift: int) -> ScheduleInfo:
+    def make_simple_schedule_and_sync(self) -> tuple[dict[str, list[list[int]]], list[SWaitCnt | SBarrier]]:
         """
-        Create a simple schedule that is valid in the mainloop, but is susceptible to race conditions in the NGL and NLL based on the value of shift.
+        Create a simple schedule that is valid in the mainloop, but is susceptible to race conditions in the NGL loop based on the value of shift.
         It is susceptible becasue only SWaitcnt for the Global Reads occurs AFTER the when the GRs in this iteration would be issued.
 
         Schedule contains 3 GRAs and 3 GRBs.
@@ -50,7 +47,6 @@ class TestValidateNglAndNll(unittest.TestCase):
             "LRA1": [[6]],
             "LRB1": [[6]]
         }
-        nglshift = nllshift = shift
 
         syncCode = [
             SWaitCnt(dscnt=0, vlcnt=-1, vscnt=-1, comment="Wait for LR0s"),
@@ -58,35 +54,32 @@ class TestValidateNglAndNll(unittest.TestCase):
             SBarrier(comment="For GRs"),
             SWaitCnt(dscnt=0, vlcnt=-1, vscnt=-1, comment="Wait for LR1s"),
         ]
-        return ScheduleInfo(1, self.num_vmfma, optSchedule, syncCode, nglshift, nllshift)
+        return optSchedule, syncCode
 
     def test_simple_case_success(self):
         """
         Test the simple schedule using the exact amount of shift required to avoid race conditions.
         """
         shift_value = 3 + 3  # 3 GRAs and 3 GRBs
-        schedule = self.make_simple_schedule(shift_value)
-        status, message = verify_lrs_and_grs(schedule, {"kernel": self.kernel}, 0)
-        assert status, f"Schedule should have passed validation but did not. {message}"
+        optSchedule, syncCode = self.make_simple_schedule_and_sync()
+        self.validate(optSchedule, syncCode, 1, shift_value, shift_value, 0, None)
 
     def test_simple_case_failure(self):
         """
         Test the simple schedule using a shift value that is too small to avoid race conditions.
         """
         shift_value = 3 + 3 - 1  # 3 GRAs and 3 GRBs - 1
-        schedule = self.make_simple_schedule(shift_value)
-        status, message = verify_lrs_and_grs(schedule, {"kernel": self.kernel}, 0)
-        assert not status, f"Schedule should have failed validation but passed. {message}"
-        assert message == "GRB at index 2 is not valid. There are no guarantees on when it will be done."
+        optSchedule, syncCode = self.make_simple_schedule_and_sync()
+        self.validate(optSchedule, syncCode, 1, shift_value, shift_value, 0,
+                                         "GRB at index 2 is not valid. There are no guarantees on when it will be done.")
 
     def test_simple_case_success_too_high(self):
         """
         Test the simple schedule using a shift value larger than needed, ensure that we don't add in negative values.
         """
         shift_value = 3 + 3 + 1  # 3 GRAs and 3 GRBs + 1
-        schedule = self.make_simple_schedule(shift_value)
-        status, message = verify_lrs_and_grs(schedule, {"kernel": self.kernel}, 0)
-        assert status, f"Schedule should have passed validation but did not. {message}"
+        optSchedule, syncCode = self.make_simple_schedule_and_sync()
+        self.validate(optSchedule, syncCode, 1, shift_value, shift_value, 0, None)
 
     def test_lr0_swait_depends_on_lr1(self):
         """
@@ -104,16 +97,13 @@ class TestValidateNglAndNll(unittest.TestCase):
             SWaitCnt(dscnt=1, vlcnt=-1, vscnt=-1, comment=""),
             SWaitCnt(dscnt=0, vlcnt=-1, vscnt=-1, comment=""),
         ]
-        sched = ScheduleInfo(1, self.num_vmfma, optSchedule, syncCode, None, None)
-        status, message = verify_lrs_and_grs(sched, {"kernel": self.kernel}, 0)
-        assert not status, f"Schedule should have failed validation but passed. {message}"
-        assert message == "Loop NLL: LRB0 at index 0 is not valid. Needed before index 6 (of next iteration), but only guaranteed at index 7."
+        self.validate(optSchedule, syncCode, 1, None, None, 0,
+                                         "Loop NLL: LRB0 at index 0 is not valid. Needed before index 6 (of next iteration), but only guaranteed at index 7.")
 
-    def test_lr0_swait_depends_on_lr1_realistic(self, useZeroDscnt = False):
+    def test_lr0_swait_depends_on_lr1_realistic(self, useZeroDscnt: bool=False):
         """
         A more realistic version of `test_lr0_swait_depends_on_lr1`. GRs are now present. Optionally checks zero dscnt option for NLL
         """
-        self.kernel = create_base_kernel()
         self.kernel["MIWaveTileA"] = 4
         self.kernel["MIWaveTileB"] = 4
         self.num_vmfma = 2 * self.kernel["MIWaveTileA"] * self.kernel["MIWaveTileB"]
@@ -152,16 +142,11 @@ class TestValidateNglAndNll(unittest.TestCase):
         }
         # We need to set nglshift and nllshift for the vlcnt adjustments
         num_gr = 2  # 2 GRs total (1 GRA + 1 GRB, but we only count the actual reads not the increments)
-        sched = ScheduleInfo(1, self.num_vmfma, optSchedule, syncTable[1::2], num_gr, num_gr, useZeroDscnt)
-        status, message = verify_lrs_and_grs(sched, {"kernel": self.kernel}, 0)
-        if useZeroDscnt:
-            assert status, f"Schedule should have passed validation but failed. {message}"
-        else:
-            assert not status, f"Schedule should have failed validation but passed. {message}"
-            assert message == "Loop NLL: LRB0 at index 3 is not valid. Needed before index 28 (of next iteration), but only guaranteed at index 31."
+        expected_error_message = None if useZeroDscnt else "Loop NLL: LRB0 at index 3 is not valid. Needed before index 28 (of next iteration), but only guaranteed at index 31."
+        self.validate(optSchedule, syncTable[1::2], 1, num_gr, num_gr, 0, expected_error_message, nllZeroDscnt=useZeroDscnt)
 
     def test_lr0_swait_depends_on_lr1_realistic_zero_dscnt(self):
         """
-        A more realistic version of `test_lr0_swait_depends_on_lr1`. GRs are now present also checks zero dscnt option for NLL
+        Same as above, but uses the zero dscnt option for NLL, so the test should now pass.
         """
-        self.test_lr0_swait_depends_on_lr1_realistic(useZeroDscnt = True)
+        self.test_lr0_swait_depends_on_lr1_realistic(useZeroDscnt=True)
